refactor realtime event page into components
